o/devicestate: move types around to fix nosectoot unit tests
